Update M7_NB1_Friends_and_Family_Both_Technical_and_Non_Technical.ipynb
M7_NB1_DavidPaul

Exercise 1
a)	An important way in which the Peer – Reward intervention in the “Friends and Family” study differs from typical gamification is that instead of trying to change the behavior of users by directly rewarding them for such changes, it attempts to change user behavior by rewarding users based on changes in behavior of their buddies. Traditional gamification methods can cause long-term damage to intrinsic motivation, while methods used in the Peer – Reward intervention attempt to help people find personal connections that are likely to maintain their effect in the longer term.

b)	One potential advantage of this design, that also included Control and Peer – View interventions, is that it enabled the researchers to isolate different incentive mechanisms related to monetary reward, the value of social information, and social pressure/influence. Another potential advantage with the Peer – Reward intervention was that the reward structure (A is rewarded for B and C’s performance, while D and E is rewarded for A’s) was that it enabled the researchers “to disambiguate and focus on the dyadic and asymmetric relationship of the person doing the activity versus the person receiving the reward, motivated to convince the first.” In other words, it enabled the researchers to study the effects of the intervention in a more natural open network setting instead of a closed system where the intervention would have been a tit-for-tat reward system (A is rewarded for B and C’s performance, while B is rewarded for A and C’s, and C is rewarded for A and B’s).